FINALLY bound data to goal element
