+++ Hypothesize/src/main.js
@@ -9,52 +9,11 @@ import Vuex from 'vuex'
 // initiate usage of vuex
 Vue.use( Vuex );
 
-// create const with new instance of Vuex to reference
-const store = new Vuex.Store({
+// import state file
+import HypothesizeStore from './Store'
 
-  //all properties and data that define our application's state
-  state:{
-    goals :
-    [
-      {
-        "title" : "Goal 1",
-        "tag"   : 'green',
-        "complete" : false
-      },
-      {
-        "title" : "Goal 2",
-        "tag"   : 'green',
-        "complete" : false
-      }
-    ]
-  },
-
-  // invoked by actions when changes to state object occur
-  mutations:{
-
-    renameTitle( state, title){
-      state.goals.title = name;
-    },
-
-    addGoal( state ){
-      let newGoal = {
-        "title" : "New Goal",
-        "tag"   : 'green',
-        "complete" : false
-      }
-    }
-  },
-
-  //can call multiple mutations or other actions and validate data before commiting to mutation
-  actions:{
-
-    changeGoalTitle({ context, title }){
-      context.commit( renameTitle, title )
-    }
-
-  }
-});
-
+// create const with new instance of Vuex store
+const store = new Vuex.Store( HypothesizeStore );
 
 Vue.config.productionTip = false
